Update README.md
Our analysis revealed that birth year, education, marital status, and income are the most significant factors in web purchases. Customers born between 1963 and 1982, with a graduate degree, and a yearly income of 53718k are the average customer profile. The last Campaign was the most successful marketing campaign, and Wines was the best-performing product. However, the underperforming channel was web purchases.